requirements.txt

streamlit
wikipedia
PyPDF2
Pillow
